Added new methods
This adds two new operations;

``mkdir_p`` and ``md5_checker``.

mkdir_p: provides the ability to create a new local directory on a system,
even if it is a nested target directory.

md5_checker: Provides the ability to check a local file to see if it has
a valid MD5sum, as provided in the method.